Moved change-background
