+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/change-background/index.tsx
