+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/presentation/dropdown/dropdown.tsx
@@ -18,7 +18,7 @@ import styled from 'styled-components'
 import { ThemeInterface } from '../../styles/styled-components'
 import Portal from '../../portal'
 import { Dropshadow } from '../../styles/mixins'
-import ChangeBackground from '../change-background'
+import ChangeBackground from '../../change-background'
 import ButtonBehavior from '../button-behavior'
 import { SFC } from '../../hoc/utils'
 import { CSSProperties } from 'react'
